requirements.txt
flask
flask-cors
python-dotenv
groq
faiss-cpu
sentence-transformers
numpy
PyMuPDF
scikit-learn